requirements.txt
discord.py
requests
python-decouple
windows-curses
google_images_search
